deploy/comparison/langgraph/requirements.txt
langgraph==0.4.1
langchain-core==0.3.51
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.4
uvloop==0.21.0
